# docs/api/snapshots/system-openapi.json
{
  "components": {
    "schemas": {
...
      "CreateLocaleRequest": {
        "properties": {
          "displayName": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "localeCode": {
            "type": "string"
          },
          "nativeName": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "LocaleView": {
        "properties": {
          "defaultLocale": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "localeCode": {
            "type": "string"
          },
          "nativeName": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ResultListLocaleView": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/LocaleView"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ResultListSupportedLocaleInfo": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/SupportedLocaleInfo"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ResultLocaleView": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/LocaleView"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ResultVoid": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {},
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "StatusRequest": {
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "SupportedLocaleInfo": {
        "properties": {
          "defaultLocale": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "localeCode": {
            "type": "string"
          },
          "nativeName": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "UpdateLocaleRequest": {
        "properties": {
          "displayName": {
            "type": "string"
          },
          "nativeName": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "VersionRequest": {
        "properties": {
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "description": "服务内部 Controller 路径；Gateway 路径与鉴权规则见 docs/api/README.md。",
    "title": "MOM System Service API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/i18n/locales": {
      "get": {
        "operationId": "supportedLocales",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListSupportedLocaleInfo"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "supported-locale-controller"
        ]
      }
    },
    "/i18n/locales/admin": {
      "get": {
        "operationId": "list",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListLocaleView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "supported-locale-controller"
        ]
      },
      "post": {
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocaleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLocaleView"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "supported-locale-controller"
        ]
      }
    },
    "/i18n/locales/admin/{id}": {
      "put": {
        "operationId": "update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocaleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLocaleView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "supported-locale-controller"
        ]
      }
    },
    "/i18n/locales/admin/{id}/default": {
      "post": {
        "operationId": "makeDefault",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLocaleView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "supported-locale-controller"
        ]
      }
    },
    "/i18n/locales/admin/{id}/status": {
      "patch": {
        "operationId": "changeStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLocaleView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "supported-locale-controller"
        ]
      }
    },
...
  }
}